requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.1